requirements.txt
openai>=1.14.0
anthropic>=0.42.0
pydantic>=2.6.0
httpx>=0.27.0
rich>=13.7.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
python-docx>=1.1.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
